update camera.far based on zoom level
However, if -- in the future -- a user creates a scene with initial camera position far away from globe without using zoom, then they won't see the globe. That is a future feature, not possible yet, but requested by users. And perhaps that's the responsibility of the camera focus function anyway, it should make sure the camera.far distance is sufficient.